+++ CHANGELOG.md
@@ -40,6 +40,7 @@ Senior Web Developer
 - **(fix)** Keep document title when HTML5 history API is enabled (nevkontakte)
 - **(fix)** Fixed wrong behavior of TbJsonGridView when doing AJAX updates like row deletion when there is pagination (nevkontakte)
 - **(enh)** Bootstrap upgrade to 2.3.1 (magefad)
+- **(fix)** Added rowHtmlOptionsExpression support for TbExtendedGridView (xt99)
 
 
 ### YiiBooster version 1.0.5